refactor: moved user actions to a separate file
