+++ src/store/reducers/user/user.action-types.ts
@@ -1,0 +1,6 @@
+const UserActionTypes = {
+  LOGIN: 'user/login',
+  LOGOUT: 'user/logout'
+}
+
+export default UserActionTypes
