+++ src/store/reducers/user/user.actions.ts
@@ -1,0 +1,9 @@
+import User from '../../../types/user.types'
+import UserActionTypes from './user.action-types'
+
+export const loginUser = (payload: User) => ({
+  type: UserActionTypes.LOGIN,
+  payload
+})
+
+export const logout = () => ({ type: UserActionTypes.LOGOUT })
