+++ src/store/reducers/user/user.reducer.ts
@@ -1,4 +1,5 @@
-import User from '../../types/user.types'
+import User from '../../../types/user.types'
+import UserActionTypes from './user.action-types'
 
 interface InitialState {
   currentUser: User | null
@@ -12,9 +13,9 @@ const inicialState: InitialState = {
 
 const userReducer = (state = inicialState, action: any) => {
   switch (action.type) {
-    case 'LOGIN_USER':
+    case UserActionTypes.LOGIN:
       return { ...state, currentUser: action.payload, isAuthenticated: true }
-    case 'LOGOUT_USER':
+    case UserActionTypes.LOGOUT:
       return { ...state, currentUser: null, isAuthenticated: false }
     default:
       return {
